Removed the check for a None object.
There is a request that requires has 1 optional attribute of the json. However, the endpoint in the api will not handle data=None but it will handle data={}.
According to a comments on the https://github.com/mitsuhiko/flask/issues/1317 None is invalid json, but an empty dict is valid.