Rewrite the documentation as a get-started guide
The `README`s for our clients should help you get set up making API
calls to Notify quickly and easily. Therefore you don’t need to know
about how to run the clients tests, or develop on it locally.

This commit rewrites the README to contain just the bare minimum you’d
need to integrate the client into a new Python project.

This is sort of a strawperson that we can talk about.