Document responses and exceptions
We don’t want the README to be complete documentation, ie all the
possible API responses (that’s what the API documentation is for, which
this commit links to).

But it should tell you what _kind_ of thing you get back from the client
(eg some JSON, a `HTTPError()`). This commit documents that.