Add helpful string representation of exceptions
It’s hard to debug what’s going on with the API when you have to know
the structure of the `Exception` object that the client returns.

This commit adds a `__str__` method to the exceptions that the client
raises. This means that the error code and message will be displayed
in any tracebacks that a developer might be looking at.