enhance setup.py to support `setup.py test` and add metadata
`python setup.py test` now works - even if you're in a completely
empty python environment, it'll download pytest and run all the tests.
if we wish to add arguments, they should be put in setup.cfg

also i added some metadata - some keyword that people might search for,
and some classifiers to flag it as MIT and python3.4 only.