Remove documentation from README
https://docs.notifications.service.gov.uk will be the canonical place
for all documentation in the future. For now only the Python client
documentation has been completely migrated.

The README of this repo is now _not_ the place for documentation. It
should just point people towards the information they’re looking for.
I’ve added the links which I think might be useful. Other suggestions
welcome.