Fix python2 document upload JSON serialization error
`open` returns a `file` object in Python 2, which isn't a subclass
of `IOBase`, so it gets sent to `json.dumps`, which raises an error.

To fix this, we need to call `.read` on subclasses of either IOBase
(which includes BytesIO and StringIO in python 2 and 3, as well as
the buffer object returned by `open` in python 3) and `file` (which
covers objects returns by `open` in python 2).

Since `file` is not defined in Python 3 this involves some attribute
testing on the `__builtins__`.

An alternative would be to check for the existence of the `.read`
method, but this could involve transforming unrelated objects into
file uploads, which otherwise would've been serialized normally.