Fix bug where we wait over 600 seconds for PDF to be ready
We shouldn't be sleeping for `count`, which sees us first sleep for 1 second,
then for 2, then for 3 etc all the way up to 35 seconds, at a total of
around 650 seconds or 11 minutes. Instead we sleep for 3 seconds, try
the API call again, sleep for 3, try again etc for up to 35 seconds as
intended.