Fix parsing of errors response
Currently the notifications client HTTPError exception will return the
errors dict for errors that look like those documented in
https://docs.notifications.service.gov.uk/rest-api.html#error-codes.

This isn't what I was expecting, and seems like a bug to me, for two
reasons:

1. the `str()` value will include the JSON in a way that looks ugly
2. the type of `error` will change depending on the format of the HTTP
response (it can be a string or a list of dicts)

This commit tries to fix this in a backwards-compatible way, by adding
two new properties, `errors` that always returns a list of dicts, and
`error_message` that always returns a `str`. `str()` of a HTTPError will
now always just include the error message, rather than the whole error
JSON object.