Add permanent-failure status for letters.
If a letter passes our validation but the print provider can not print
the letter we should mark it as permanent-failure. An error message for
permanent-failure is being added to the UI.
Technical-failure had the wrong message and validation failure will not
show the pdf because the letter is in the wrong bucket.