Support encoding sets in JSON POST data
https://www.pivotaltracker.com/story/show/178770155

Sets are just a special kind of list, so there's no reason why we
can't support them in POST data. Rather than peppering our code with
list(set) conversions (and testing them), it seems more pragmatic to
do it once where we can get the most benefit from it.

One use case for this will be POSTing permissions from Admin to API.
Admin uses sets for permissions most of the time, which means we get
an error if we try to POST them directly:

      File "/Users/benthorner/.pyenv/versions/3.6.12/lib/python3.6/json/encoder.py", line 180, in default
        o.__class__.__name__)
      TypeError: Object of type 'set' is not JSON serializable