Emit a warning rather than an error
When the API call results in a RequestError, let's only log out a
warning rather than an error-level log line as the request failure could
be benign (eg a reasonable 404). We still throw the exception up the
chain so any calling code can log out/handle an error if needed.