install requirements.in without pinned subdependencies
our subdependencies are pinned when we run `make bootstrap`,
which happens with the primary pyenv version of python (currently 3.12).

However, this results in a set of subdependencies which is only
guaranteed to be valid for python 3.12 itself. This has resulted in
trying to install a version of coverage that isn't compatible with 3.8,
and because every single subdep is pinned, the pip resolver has no
flexibility even if the dependency (pytest-testmon) in this case has a
much wider range of allowed coverage versions.

So, for now, just install `requirements.in` (ie: the dependencies we
actually use).

Lets ignore the fact that 90% of these modules we don't actually use in
the python-client itself