Dockerfile: Install awscli & jq
In the new deploy pipeline, this will be used to fetch the SSM parameter
containing the environment variables required to run the integration
tests.

Also removes the comment explaining the `COPY . .` line, because now we
don't only need these files in order to run `make bootstrap`, we need
them to be able to run the tests at all without packing the repo into
the deployment bag.